refactor(semver): one parser, reachable from node as well as bun
`src/semver.ts` unified three copies but left four more, because the release
scripts that also parse versions run under `node` in npm-publish.yml,
release-npm-publish.yml and homebrew-tap.yml — and node cannot import a `.ts`
module. So the module becomes `src/semver.mjs` with types in `semver.d.mts`,
the arrangement `npm-dist-tag.d.mts` already uses for the same reason.

Folded in:

- `set-package-version.mjs` had its own SEMVER_RE, looser than the gate's: it
  would have written `01.2.3` into package.json.
- `npm-dist-tag.mjs` sliced the channel out of the string instead of parsing.
  That read `not-a-version` as the channel `a-version` and published to it;
  now the grammar rejects it. `1.27.0-.1` and `1.27.0-` move from "not a usable
  channel name" to "not a version", which is what they are — both still refuse.
- `linux-package-names.mjs` and `update-homebrew-tap.mjs` each carried their own
  copy of "stable X.Y.Z"; the first uses `isStableVersion`, the second the
  `isStableTag` that already existed next to it.
- `src/star.ts` counted dot-separated parts to find major/minor.

`tryParseSemver` is the only place the pattern is applied; `parseSemver`,
`isSemver` and `isStableVersion` build on it, so no caller reaches for a regex
or a try/catch to ask whether a string is a version.

Deliberately left alone: `release-tags.mjs` declares the release *tag* grammar
as a POSIX ERE string so bash `=~` in build-engine.yml and JavaScript consume
identical text, pinned by release-manifest.mjs (#685). A JavaScript parser
cannot be read by bash, so folding it in would break what it exists to protect.
`upload-flakiness.ts` parses macOS versions, which are not SemVer.

Every touched script exercised under both node and bun.

Refs #738, #685
